refactor signing info test
